% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.databasemigrationservice_operations.R
\name{create_endpoint}
\alias{create_endpoint}
\title{Creates an endpoint using the provided settings}
\usage{
create_endpoint(EndpointIdentifier, EndpointType, EngineName,
  Username = NULL, Password = NULL, ServerName = NULL, Port = NULL,
  DatabaseName = NULL, ExtraConnectionAttributes = NULL,
  KmsKeyId = NULL, Tags = NULL, CertificateArn = NULL,
  SslMode = NULL, ServiceAccessRoleArn = NULL,
  ExternalTableDefinition = NULL, DynamoDbSettings = NULL,
  S3Settings = NULL, DmsTransferSettings = NULL,
  MongoDbSettings = NULL)
}
\arguments{
\item{EndpointIdentifier}{[required] The database endpoint identifier. Identifiers must begin with a letter; must contain only ASCII letters, digits, and hyphens; and must not end with a hyphen or contain two consecutive hyphens.}

\item{EndpointType}{[required] The type of endpoint.}

\item{EngineName}{[required] The type of engine for the endpoint. Valid values, depending on the EndPointType, include mysql, oracle, postgres, mariadb, aurora, aurora-postgresql, redshift, s3, db2, azuredb, sybase, dynamodb, mongodb, and sqlserver.}

\item{Username}{The user name to be used to login to the endpoint database.}

\item{Password}{The password to be used to login to the endpoint database.}

\item{ServerName}{The name of the server where the endpoint database resides.}

\item{Port}{The port used by the endpoint database.}

\item{DatabaseName}{The name of the endpoint database.}

\item{ExtraConnectionAttributes}{Additional attributes associated with the connection.}

\item{KmsKeyId}{The KMS key identifier that will be used to encrypt the connection parameters. If you do not specify a value for the KmsKeyId parameter, then AWS DMS will use your default encryption key. AWS KMS creates the default encryption key for your AWS account. Your AWS account has a different default encryption key for each AWS region.}

\item{Tags}{Tags to be added to the endpoint.}

\item{CertificateArn}{The Amazon Resource Name (ARN) for the certificate.}

\item{SslMode}{The SSL mode to use for the SSL connection.

SSL mode can be one of four values: none, require, verify-ca, verify-full.

The default value is none.}

\item{ServiceAccessRoleArn}{The Amazon Resource Name (ARN) for the service access role you want to use to create the endpoint.}

\item{ExternalTableDefinition}{The external table definition.}

\item{DynamoDbSettings}{Settings in JSON format for the target Amazon DynamoDB endpoint. For more information about the available settings, see the \strong{Using Object Mapping to Migrate Data to DynamoDB} section at \href{http://docs.aws.amazon.com/dms/latest/userguide/CHAP_Target.DynamoDB.html}{Using an Amazon DynamoDB Database as a Target for AWS Database Migration Service}.}

\item{S3Settings}{Settings in JSON format for the target Amazon S3 endpoint. For more information about the available settings, see the \strong{Extra Connection Attributes} section at \href{http://docs.aws.amazon.com/dms/latest/userguide/CHAP_Target.S3.html}{Using Amazon S3 as a Target for AWS Database Migration Service}.}

\item{DmsTransferSettings}{The settings in JSON format for the DMS Transfer type source endpoint.

Attributes include:
\itemize{
\item serviceAccessRoleArn - The IAM role that has permission to access the Amazon S3 bucket.
\item bucketName - The name of the S3 bucket to use.
\item compressionType - An optional parameter to use GZIP to compress the target files. Set to NONE (the default) or do not use to leave the files uncompressed.
}

Shorthand syntax: ServiceAccessRoleArn=string ,BucketName=string,CompressionType=string

JSON syntax:

{ "ServiceAccessRoleArn": "string", "BucketName": "string", "CompressionType": "none"|"gzip" }}

\item{MongoDbSettings}{Settings in JSON format for the source MongoDB endpoint. For more information about the available settings, see the \strong{Configuration Properties When Using MongoDB as a Source for AWS Database Migration Service} section at \href{http://docs.aws.amazon.com/dms/latest/userguide/CHAP_Source.MongoDB.html}{Using MongoDB as a Target for AWS Database Migration Service}.}
}
\description{
Creates an endpoint using the provided settings.
}
\section{Accepted Parameters}{
\preformatted{create_endpoint(
  EndpointIdentifier = "string",
  EndpointType = "source"|"target",
  EngineName = "string",
  Username = "string",
  Password = "string",
  ServerName = "string",
  Port = 123,
  DatabaseName = "string",
  ExtraConnectionAttributes = "string",
  KmsKeyId = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  CertificateArn = "string",
  SslMode = "none"|"require"|"verify-ca"|"verify-full",
  ServiceAccessRoleArn = "string",
  ExternalTableDefinition = "string",
  DynamoDbSettings = list(
    ServiceAccessRoleArn = "string"
  ),
  S3Settings = list(
    ServiceAccessRoleArn = "string",
    ExternalTableDefinition = "string",
    CsvRowDelimiter = "string",
    CsvDelimiter = "string",
    BucketFolder = "string",
    BucketName = "string",
    CompressionType = "none"|"gzip"
  ),
  DmsTransferSettings = list(
    ServiceAccessRoleArn = "string",
    BucketName = "string"
  ),
  MongoDbSettings = list(
    Username = "string",
    Password = "string",
    ServerName = "string",
    Port = 123,
    DatabaseName = "string",
    AuthType = "no"|"password",
    AuthMechanism = "default"|"mongodb_cr"|"scram_sha_1",
    NestingLevel = "none"|"one",
    ExtractDocId = "string",
    DocsToInvestigate = "string",
    AuthSource = "string",
    KmsKeyId = "string"
  )
)
}
}

\examples{
# Creates an endpoint using the provided settings.
\donttest{create_endpoint(
  CertificateArn = "",
  DatabaseName = "testdb",
  EndpointIdentifier = "test-endpoint-1",
  EndpointType = "source",
  EngineName = "mysql",
  ExtraConnectionAttributes = "",
  KmsKeyId = "arn:aws:kms:us-east-1:123456789012:key/4c1731d6-5435-ed4d-be13-d53411a7cfbd",
  Password = "pasword",
  Port = 3306L,
  ServerName = "mydb.cx1llnox7iyx.us-west-2.rds.amazonaws.com",
  SslMode = "require",
  Tags = list(
    list(
      Key = "Acount",
      Value = "143327655"
    )
  ),
  Username = "username"
)}

}
