% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.databasemigrationservice_operations.R
\name{create_replication_task}
\alias{create_replication_task}
\title{Creates a replication task using the specified parameters}
\usage{
create_replication_task(
  ReplicationTaskIdentifier = "string",
  SourceEndpointArn = "string",
  TargetEndpointArn = "string",
  ReplicationInstanceArn = "string",
  MigrationType = "full-load"|"cdc"|"full-load-and-cdc",
  TableMappings = "string",
  ReplicationTaskSettings = "string",
  CdcStartTime = as.POSIXct("2015-01-01"),
  CdcStartPosition = "string",
  CdcStopPosition = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
\arguments{
\item{ReplicationTaskIdentifier}{[required] The replication task identifier.

Constraints:
\itemize{
\item Must contain from 1 to 255 alphanumeric characters or hyphens.
\item First character must be a letter.
\item Cannot end with a hyphen or contain two consecutive hyphens.
}}

\item{SourceEndpointArn}{[required] The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.}

\item{TargetEndpointArn}{[required] The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.}

\item{ReplicationInstanceArn}{[required] The Amazon Resource Name (ARN) of the replication instance.}

\item{MigrationType}{[required] The migration type.}

\item{TableMappings}{[required] When using the AWS CLI or boto3, provide the path of the JSON file that contains the table mappings. Precede the path with "file://". When working with the DMS API, provide the JSON as the parameter value.

For example, --table-mappings file://mappingfile.json}

\item{ReplicationTaskSettings}{Settings for the task, such as target metadata settings. For a complete list of task settings, see \href{http://docs.aws.amazon.com/dms/latest/userguide/CHAP_Tasks.CustomizingTasks.TaskSettings.html}{Task Settings for AWS Database Migration Service Tasks} in the \emph{AWS Database Migration User Guide.}}

\item{CdcStartTime}{Indicates the start time for a change data capture (CDC) operation. Use either CdcStartTime or CdcStartPosition to specify when you want a CDC operation to start. Specifying both values results in an error.

Timestamp Example: --cdc-start-time "2018-03-08T12:12:12"}

\item{CdcStartPosition}{Indicates when you want a change data capture (CDC) operation to start. Use either CdcStartPosition or CdcStartTime to specify when you want a CDC operation to start. Specifying both values results in an error.

The value can be in date, checkpoint, or LSN/SCN format.

Date Example: --cdc-start-position "2018-03-08T12:12:12"

Checkpoint Example: --cdc-start-position "checkpoint:V1#27#mysql-bin-changelog.157832:1975:-1:2002:677883278264080:mysql-bin-changelog.157832:1876#0#0#*#0#93"

LSN Example: --cdc-start-position "mysql-bin-changelog.000024:373"}

\item{CdcStopPosition}{Indicates when you want a change data capture (CDC) operation to stop. The value can be either server time or commit time.

Server time example: --cdc-stop-position "server\_time:3018-02-09T12:12:12"

Commit time example: --cdc-stop-position "commit\_time: 3018-02-09T12:12:12 "}

\item{Tags}{Tags to be added to the replication instance.}
}
\description{
Creates a replication task using the specified parameters.
}
\examples{
# Creates a replication task using the specified parameters.
create_replication_task(
  CdcStartTime = "2016-12-14T18:25:43Z",
  MigrationType = "full-load",
  ReplicationInstanceArn = "arn:aws:dms:us-east-1:123456789012:rep:6UTDJGBOUS3VI3SUWA66XFJCJQ",
  ReplicationTaskIdentifier = "task1",
  ReplicationTaskSettings = "",
  SourceEndpointArn = "arn:aws:dms:us-east-1:123456789012:endpoint:ZW5UAN6P4E77EC7YWHK4RZZ3BE",
  TableMappings = "file://mappingfile.json",
  Tags = list(
    list(
      Key = "Acount",
      Value = "24352226"
    )
  ),
  TargetEndpointArn = "arn:aws:dms:us-east-1:123456789012:endpoint:ASXWXJZLNWNT5HTWCGV2BUJQ7E"
)

}
