% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.databasemigrationservice_operations.R
\name{delete_endpoint}
\alias{delete_endpoint}
\title{Deletes the specified endpoint}
\usage{
delete_endpoint(EndpointArn)
}
\arguments{
\item{EndpointArn}{[required] The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.}
}
\description{
Deletes the specified endpoint.
}
\details{
All tasks associated with the endpoint must be deleted before you can delete the endpoint.
}
\section{Accepted Parameters}{
\preformatted{delete_endpoint(
  EndpointArn = "string"
)
}
}

\examples{
# Deletes the specified endpoint. All tasks associated with the endpoint
# must be deleted before you can delete the endpoint.
# 
\donttest{delete_endpoint(
  EndpointArn = "arn:aws:dms:us-east-1:123456789012:endpoint:RAAR3R22XSH46S3PWLC3NJAWKM"
)}

}
