% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.databasemigrationservice_operations.R
\name{delete_event_subscription}
\alias{delete_event_subscription}
\title{Deletes an AWS DMS event subscription}
\usage{
delete_event_subscription(SubscriptionName)
}
\arguments{
\item{SubscriptionName}{The name of the DMS event notification subscription to be deleted.}
}
\description{
Deletes an AWS DMS event subscription.
}
\examples{

}
