% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.databasemigrationservice_operations.R
\name{delete_replication_instance}
\alias{delete_replication_instance}
\title{Deletes the specified replication instance}
\usage{
delete_replication_instance(ReplicationInstanceArn)
}
\arguments{
\item{ReplicationInstanceArn}{The Amazon Resource Name (ARN) of the replication instance to be deleted.}
}
\description{
Deletes the specified replication instance.
}
\details{
You must delete any migration tasks that are associated with the replication instance before you can delete it.
}
\examples{
# Deletes the specified replication instance. You must delete any
# migration tasks that are associated with the replication instance before
# you can delete it.
# 
# 
delete_replication_instance(
  ReplicationInstanceArn = "arn:aws:dms:us-east-1:123456789012:rep:6UTDJGBOUS3VI3SUWA66XFJCJQ"
)

}
