% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.databasemigrationservice_operations.R
\name{delete_replication_subnet_group}
\alias{delete_replication_subnet_group}
\title{Deletes a subnet group}
\usage{
delete_replication_subnet_group(ReplicationSubnetGroupIdentifier)
}
\arguments{
\item{ReplicationSubnetGroupIdentifier}{The subnet group name of the replication instance.}
}
\description{
Deletes a subnet group.
}
\examples{
# Deletes a replication subnet group.
delete_replication_subnet_group(
  ReplicationSubnetGroupIdentifier = "us-west-2ab-vpc-215ds366"
)

}
