% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.databasemigrationservice_operations.R
\name{describe_events}
\alias{describe_events}
\title{Lists events for a given source identifier and source type}
\usage{
describe_events(SourceIdentifier = NULL, SourceType = NULL,
  StartTime = NULL, EndTime = NULL, Duration = NULL,
  EventCategories = NULL, Filters = NULL, MaxRecords = NULL,
  Marker = NULL)
}
\arguments{
\item{SourceIdentifier}{The identifier of the event source. An identifier must begin with a letter and must contain only ASCII letters, digits, and hyphens. It cannot end with a hyphen or contain two consecutive hyphens.}

\item{SourceType}{The type of AWS DMS resource that generates events.

Valid values: replication-instance | migration-task}

\item{StartTime}{The start time for the events to be listed.}

\item{EndTime}{The end time for the events to be listed.}

\item{Duration}{The duration of the events to be listed.}

\item{EventCategories}{A list of event categories for a source type that you want to subscribe to.}

\item{Filters}{Filters applied to the action.}

\item{MaxRecords}{The maximum number of records to include in the response. If more records exist than the specified \code{MaxRecords} value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.}

\item{Marker}{An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by \code{MaxRecords}.}
}
\description{
Lists events for a given source identifier and source type. You can also specify a start and end time. For more information on AWS DMS events, see \href{http://docs.aws.amazon.com/dms/latest/userguide/CHAP_Events.html}{Working with Events and Notifications} in the \emph{AWS Database Migration User Guide.}
}
\section{Accepted Parameters}{
\preformatted{describe_events(
  SourceIdentifier = "string",
  SourceType = "replication-instance",
  StartTime = as.POSIXct("2015-01-01"),
  EndTime = as.POSIXct("2015-01-01"),
  Duration = 123,
  EventCategories = list(
    "string"
  ),
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  ),
  MaxRecords = 123,
  Marker = "string"
)
}
}

