% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.databasemigrationservice_operations.R
\name{describe_refresh_schemas_status}
\alias{describe_refresh_schemas_status}
\title{Returns the status of the RefreshSchemas operation}
\usage{
describe_refresh_schemas_status(EndpointArn)
}
\arguments{
\item{EndpointArn}{[required] The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.}
}
\description{
Returns the status of the RefreshSchemas operation.
}
\section{Accepted Parameters}{
\preformatted{describe_refresh_schemas_status(
  EndpointArn = "string"
)
}
}

\examples{
# Returns the status of the refresh-schemas operation.
\donttest{describe_refresh_schemas_status(
  EndpointArn = ""
)}

}
