% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.databasemigrationservice_operations.R
\name{describe_replication_instance_task_logs}
\alias{describe_replication_instance_task_logs}
\title{Returns information about the task logs for the specified task}
\usage{
describe_replication_instance_task_logs(
  ReplicationInstanceArn = "string",
  MaxRecords = 123,
  Marker = "string"
)
}
\arguments{
\item{ReplicationInstanceArn}{[required] The Amazon Resource Name (ARN) of the replication instance.}

\item{MaxRecords}{The maximum number of records to include in the response. If more records exist than the specified \code{MaxRecords} value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.}

\item{Marker}{An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by \code{MaxRecords}.}
}
\description{
Returns information about the task logs for the specified task.
}
\examples{

}
