% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.databasemigrationservice_operations.R
\name{describe_replication_subnet_groups}
\alias{describe_replication_subnet_groups}
\title{Returns information about the replication subnet groups}
\usage{
describe_replication_subnet_groups(Filters = NULL, MaxRecords = NULL,
  Marker = NULL)
}
\arguments{
\item{Filters}{Filters applied to the describe action.}

\item{MaxRecords}{The maximum number of records to include in the response. If more records exist than the specified \code{MaxRecords} value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.}

\item{Marker}{An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by \code{MaxRecords}.}
}
\description{
Returns information about the replication subnet groups.
}
\section{Accepted Parameters}{
\preformatted{describe_replication_subnet_groups(
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string"
      )
    )
  ),
  MaxRecords = 123,
  Marker = "string"
)
}
}

\examples{
# Returns information about the replication subnet groups.
\donttest{describe_replication_subnet_groups(
  Filters = list(
    list(
      Name = "string",
      Values = list(
        "string",
        "string"
      )
    )
  ),
  Marker = "",
  MaxRecords = 123L
)}

}
