% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.databasemigrationservice_operations.R
\name{describe_replication_task_assessment_results}
\alias{describe_replication_task_assessment_results}
\title{Returns the task assessment results from Amazon S3}
\usage{
describe_replication_task_assessment_results(ReplicationTaskArn = NULL,
  MaxRecords = NULL, Marker = NULL)
}
\arguments{
\item{ReplicationTaskArn}{- The Amazon Resource Name (ARN) string that uniquely identifies the task. When this input parameter is specified the API will return only one result and ignore the values of the max-records and marker parameters.}

\item{MaxRecords}{The maximum number of records to include in the response. If more records exist than the specified \code{MaxRecords} value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.}

\item{Marker}{An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by \code{MaxRecords}.}
}
\description{
Returns the task assessment results from Amazon S3. This action always returns the latest results.
}
\section{Accepted Parameters}{
\preformatted{describe_replication_task_assessment_results(
  ReplicationTaskArn = "string",
  MaxRecords = 123,
  Marker = "string"
)
}
}

