% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.databasemigrationservice_operations.R
\name{import_certificate}
\alias{import_certificate}
\title{Uploads the specified certificate}
\usage{
import_certificate(
  CertificateIdentifier = "string",
  CertificatePem = "string",
  CertificateWallet = raw,
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
\arguments{
\item{CertificateIdentifier}{[required] The customer-assigned name of the certificate. Valid characters are A-z and 0-9.}

\item{CertificatePem}{The contents of the .pem X.509 certificate file for the certificate.}

\item{CertificateWallet}{The location of the imported Oracle Wallet certificate for use with SSL.}

\item{Tags}{The tags associated with the certificate.}
}
\description{
Uploads the specified certificate.
}
\examples{
# Uploads the specified certificate.
import_certificate(
  CertificateIdentifier = "",
  CertificatePem = ""
)

}
