% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.databasemigrationservice_operations.R
\name{list_tags_for_resource}
\alias{list_tags_for_resource}
\title{Lists all tags for an AWS DMS resource}
\usage{
list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{The Amazon Resource Name (ARN) string that uniquely identifies the AWS DMS resource.}
}
\description{
Lists all tags for an AWS DMS resource.
}
\examples{
# Lists all tags for an AWS DMS resource.
list_tags_for_resource(
  ResourceArn = ""
)

}
