% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.databasemigrationservice_operations.R
\name{modify_replication_subnet_group}
\alias{modify_replication_subnet_group}
\title{Modifies the settings for the specified replication subnet group}
\usage{
modify_replication_subnet_group(ReplicationSubnetGroupIdentifier, SubnetIds,
  ReplicationSubnetGroupDescription = NULL)
}
\arguments{
\item{ReplicationSubnetGroupIdentifier}{The name of the replication instance subnet group.}

\item{SubnetIds}{A list of subnet IDs.}

\item{ReplicationSubnetGroupDescription}{The description of the replication instance subnet group.}
}
\description{
Modifies the settings for the specified replication subnet group.
}
\examples{
# Modifies the settings for the specified replication subnet group.
modify_replication_subnet_group(
  ReplicationSubnetGroupDescription = "",
  ReplicationSubnetGroupIdentifier = "",
  SubnetIds = list()
)

}
