% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.databasemigrationservice_operations.R
\name{reboot_replication_instance}
\alias{reboot_replication_instance}
\title{Reboots a replication instance}
\usage{
reboot_replication_instance(ReplicationInstanceArn, ForceFailover = NULL)
}
\arguments{
\item{ReplicationInstanceArn}{[required] The Amazon Resource Name (ARN) of the replication instance.}

\item{ForceFailover}{If this parameter is \code{true}, the reboot is conducted through a Multi-AZ failover. (If the instance isn't configured for Multi-AZ, then you can't specify \code{true}.)}
}
\description{
Reboots a replication instance. Rebooting results in a momentary outage, until the replication instance becomes available again.
}
\section{Accepted Parameters}{
\preformatted{reboot_replication_instance(
  ReplicationInstanceArn = "string",
  ForceFailover = TRUE|FALSE
)
}
}

