% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.databasemigrationservice_operations.R
\name{remove_tags_from_resource}
\alias{remove_tags_from_resource}
\title{Removes metadata tags from a DMS resource}
\usage{
remove_tags_from_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{>The Amazon Resource Name (ARN) of the AWS DMS resource the tag is to be removed from.}

\item{TagKeys}{The tag key (name) of the tag to be removed.}
}
\description{
Removes metadata tags from a DMS resource.
}
\examples{
# Removes metadata tags from an AWS DMS resource.
remove_tags_from_resource(
  ResourceArn = "arn:aws:dms:us-east-1:123456789012:endpoint:ASXWXJZLNWNT5HTWCGV2BUJQ7E",
  TagKeys = list()
)

}
