% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.databasemigrationservice_operations.R
\name{start_replication_task_assessment}
\alias{start_replication_task_assessment}
\title{Starts the replication task assessment for unsupported data types in the source database}
\usage{
start_replication_task_assessment(ReplicationTaskArn)
}
\arguments{
\item{ReplicationTaskArn}{The Amazon Resource Name (ARN) of the replication task.}
}
\description{
Starts the replication task assessment for unsupported data types in the source database.
}
\examples{

}
