% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.databasemigrationservice_operations.R
\name{stop_replication_task}
\alias{stop_replication_task}
\title{Stops the replication task}
\usage{
stop_replication_task(ReplicationTaskArn)
}
\arguments{
\item{ReplicationTaskArn}{[required] The Amazon Resource Name(ARN) of the replication task to be stopped.}
}
\description{
Stops the replication task.
}
\section{Accepted Parameters}{
\preformatted{stop_replication_task(
  ReplicationTaskArn = "string"
)
}
}

\examples{
# Stops the replication task.
\donttest{stop_replication_task(
  ReplicationTaskArn = "arn:aws:dms:us-east-1:123456789012:endpoint:ASXWXJZLNWNT5HTWCGV2BUJQ7E"
)}

}
