% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.databasemigrationservice_operations.R
\name{test_connection}
\alias{test_connection}
\title{Tests the connection between the replication instance and the endpoint}
\usage{
test_connection(ReplicationInstanceArn, EndpointArn)
}
\arguments{
\item{ReplicationInstanceArn}{The Amazon Resource Name (ARN) of the replication instance.}

\item{EndpointArn}{The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.}
}
\description{
Tests the connection between the replication instance and the endpoint.
}
\examples{
# Tests the connection between the replication instance and the endpoint.
test_connection(
  EndpointArn = "arn:aws:dms:us-east-1:123456789012:endpoint:RAAR3R22XSH46S3PWLC3NJAWKM",
  ReplicationInstanceArn = "arn:aws:dms:us-east-1:123456789012:rep:6UTDJGBOUS3VI3SUWA66XFJCJQ"
)

}
