% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.databasemigrationservice_operations.R
\name{delete_replication_task}
\alias{delete_replication_task}
\title{Deletes the specified replication task}
\usage{
delete_replication_task(
  ReplicationTaskArn = "string"
)
}
\arguments{
\item{ReplicationTaskArn}{[required] The Amazon Resource Name (ARN) of the replication task to be deleted.}
}
\description{
Deletes the specified replication task.
}
\examples{
# Deletes the specified replication task.
delete_replication_task(
  ReplicationTaskArn = "arn:aws:dms:us-east-1:123456789012:rep:6UTDJGBOUS3VI3SUWA66XFJCJQ"
)

}
