% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.databasemigrationservice_operations.R
\name{describe_account_attributes}
\alias{describe_account_attributes}
\title{Lists all of the AWS DMS attributes for a customer account}
\usage{
describe_account_attributes()
}
\description{
Lists all of the AWS DMS attributes for a customer account. The attributes include AWS DMS quotas for the account, such as the number of replication instances allowed. The description for a quota includes the quota name, current usage toward that quota, and the quota's maximum value.
}
\details{
This command does not take any parameters.
}
\examples{
# Lists all of the AWS DMS attributes for a customer account. The
# attributes include AWS DMS quotas for the account, such as the number of
# replication instances allowed. The description for a quota includes the
# quota name, current usage toward that quota, and the quota's maximum
# value. This operation does not take any parameters.
describe_account_attributes()

}
