% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.databasemigrationservice_operations.R
\name{modify_endpoint}
\alias{modify_endpoint}
\title{Modifies the specified endpoint}
\usage{
modify_endpoint(EndpointArn, EndpointIdentifier = NULL,
  EndpointType = NULL, EngineName = NULL, Username = NULL,
  Password = NULL, ServerName = NULL, Port = NULL,
  DatabaseName = NULL, ExtraConnectionAttributes = NULL,
  CertificateArn = NULL, SslMode = NULL, ServiceAccessRoleArn = NULL,
  ExternalTableDefinition = NULL, DynamoDbSettings = NULL,
  S3Settings = NULL, DmsTransferSettings = NULL,
  MongoDbSettings = NULL)
}
\arguments{
\item{EndpointArn}{[required] The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.}

\item{EndpointIdentifier}{The database endpoint identifier. Identifiers must begin with a letter; must contain only ASCII letters, digits, and hyphens; and must not end with a hyphen or contain two consecutive hyphens.}

\item{EndpointType}{The type of endpoint.}

\item{EngineName}{The type of engine for the endpoint. Valid values, depending on the EndPointType, include mysql, oracle, postgres, mariadb, aurora, aurora-postgresql, redshift, s3, db2, azuredb, sybase, sybase, dynamodb, mongodb, and sqlserver.}

\item{Username}{The user name to be used to login to the endpoint database.}

\item{Password}{The password to be used to login to the endpoint database.}

\item{ServerName}{The name of the server where the endpoint database resides.}

\item{Port}{The port used by the endpoint database.}

\item{DatabaseName}{The name of the endpoint database.}

\item{ExtraConnectionAttributes}{Additional attributes associated with the connection. To reset this parameter, pass the empty string ("") as an argument.}

\item{CertificateArn}{The Amazon Resource Name (ARN) of the certificate used for SSL connection.}

\item{SslMode}{The SSL mode to be used.

SSL mode can be one of four values: none, require, verify-ca, verify-full.

The default value is none.}

\item{ServiceAccessRoleArn}{The Amazon Resource Name (ARN) for the service access role you want to use to modify the endpoint.}

\item{ExternalTableDefinition}{The external table definition.}

\item{DynamoDbSettings}{Settings in JSON format for the target Amazon DynamoDB endpoint. For more information about the available settings, see the \strong{Using Object Mapping to Migrate Data to DynamoDB} section at \href{http://docs.aws.amazon.com/dms/latest/userguide/CHAP_Target.DynamoDB.html}{Using an Amazon DynamoDB Database as a Target for AWS Database Migration Service}.}

\item{S3Settings}{Settings in JSON format for the target S3 endpoint. For more information about the available settings, see the \strong{Extra Connection Attributes} section at \href{http://docs.aws.amazon.com/dms/latest/userguide/CHAP_Target.S3.html}{Using Amazon S3 as a Target for AWS Database Migration Service}.}

\item{DmsTransferSettings}{The settings in JSON format for the DMS Transfer type source endpoint.

Attributes include:
\itemize{
\item serviceAccessRoleArn - The IAM role that has permission to access the Amazon S3 bucket.
\item BucketName - The name of the S3 bucket to use.
\item compressionType - An optional parameter to use GZIP to compress the target files. Set to NONE (the default) or do not use to leave the files uncompressed.
}

Shorthand syntax: ServiceAccessRoleArn=string ,BucketName=string,CompressionType=string

JSON syntax:

{ "ServiceAccessRoleArn": "string", "BucketName": "string", "CompressionType": "none"|"gzip" }}

\item{MongoDbSettings}{Settings in JSON format for the source MongoDB endpoint. For more information about the available settings, see the \strong{Configuration Properties When Using MongoDB as a Source for AWS Database Migration Service} section at \href{http://docs.aws.amazon.com/dms/latest/userguide/CHAP_Source.MongoDB.html}{Using Amazon S3 as a Target for AWS Database Migration Service}.}
}
\description{
Modifies the specified endpoint.
}
\section{Accepted Parameters}{
\preformatted{modify_endpoint(
  EndpointArn = "string",
  EndpointIdentifier = "string",
  EndpointType = "source"|"target",
  EngineName = "string",
  Username = "string",
  Password = "string",
  ServerName = "string",
  Port = 123,
  DatabaseName = "string",
  ExtraConnectionAttributes = "string",
  CertificateArn = "string",
  SslMode = "none"|"require"|"verify-ca"|"verify-full",
  ServiceAccessRoleArn = "string",
  ExternalTableDefinition = "string",
  DynamoDbSettings = list(
    ServiceAccessRoleArn = "string"
  ),
  S3Settings = list(
    ServiceAccessRoleArn = "string",
    ExternalTableDefinition = "string",
    CsvRowDelimiter = "string",
    CsvDelimiter = "string",
    BucketFolder = "string",
    BucketName = "string",
    CompressionType = "none"|"gzip"
  ),
  DmsTransferSettings = list(
    ServiceAccessRoleArn = "string",
    BucketName = "string"
  ),
  MongoDbSettings = list(
    Username = "string",
    Password = "string",
    ServerName = "string",
    Port = 123,
    DatabaseName = "string",
    AuthType = "no"|"password",
    AuthMechanism = "default"|"mongodb_cr"|"scram_sha_1",
    NestingLevel = "none"|"one",
    ExtractDocId = "string",
    DocsToInvestigate = "string",
    AuthSource = "string",
    KmsKeyId = "string"
  )
)
}
}

\examples{
# Modifies the specified endpoint.
\donttest{modify_endpoint(
  CertificateArn = "",
  DatabaseName = "",
  EndpointArn = "",
  EndpointIdentifier = "",
  EndpointType = "source",
  EngineName = "",
  ExtraConnectionAttributes = "",
  Password = "",
  Port = 123L,
  ServerName = "",
  SslMode = "require",
  Username = ""
)}

}
